# app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Flask REST API",
    "version": "1.0.0",
    "description": "API untuk manajemen pengguna dan stok barang menggunakan Flask dan SQLite."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000/api",
      "description": "Server Lokal"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "UserRegister": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
            "type": "string",
            "example": "user1"
          },
          "password": {
            "type": "string",
            "example": "pass123",
            "minLength": 6
          }
        }
      },
      "UserLogin": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
            "type": "string",
            "example": "user1"
          },
          "password": {
            "type": "string",
            "example": "pass123"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "username": {
            "type": "string",
            "example": "user1"
          }
        }
      },
      "UserUpdate": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "new_username"
          },
          "password": {
            "type": "string",
            "example": "newpass123",
            "minLength": 6
          }
        }
      },
      "ResetPassword": {
        "type": "object",
        "required": ["new_password"],
        "properties": {
          "new_password": {
            "type": "string",
            "example": "newpass123",
            "minLength": 6
          }
        }
      },
      "AuthToken": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
          }
        }
      },
      "StockAdd": {
        "type": "object",
        "required": ["item_name", "quantity"],
        "properties": {
          "item_name": {
            "type": "string",
            "example": "Barang A"
          },
          "quantity": {
            "type": "integer",
            "example": 50,
            "minimum": 0
          }
        }
      },
      "StockRemove": {
        "type": "object",
        "required": ["item_name", "quantity"],
        "properties": {
          "item_name": {
            "type": "string",
            "example": "Barang A"
          },
          "quantity": {
            "type": "integer",
            "example": 5,
            "minimum": 0
          }
        }
      },
      "StockResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "item_name": {
            "type": "string",
            "example": "Barang A"
          },
          "quantity": {
            "type": "integer",
            "example": 45
          }
        }
      },
      "StockDelete": {
        "type": "object",
        "required": ["item_name"],
        "properties": {
          "item_name": {
            "type": "string",
            "example": "Barang A"
          }
        }
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Operation successful."
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 400
          },
          "name": {
            "type": "string",
            "example": "Bad Request"
          },
          "description": {
            "type": "string",
            "example": "Detailed error message."
          }
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/users/register": {
      "post": {
        "summary": "Registrasi Pengguna Baru",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegister"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pengguna berhasil dibuat.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Permintaan tidak valid atau pengguna sudah ada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "Login Pengguna",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login berhasil, token akses diberikan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthToken"
                }
              }
            }
          },
          "400": {
            "description": "Data login tidak lengkap.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Kredensial tidak valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/logout": {
      "post": {
        "summary": "Logout Pengguna",
        "tags": ["Users"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logout berhasil.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token akses tidak valid atau tidak diberikan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/reset_password": {
      "post": {
        "summary": "Reset Password Pengguna",
        "tags": ["Users"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPassword"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password berhasil direset.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Data tidak valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token akses tidak valid atau tidak diberikan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Pengguna tidak ditemukan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/delete_user": {
      "delete": {
        "summary": "Hapus Akun Pengguna",
        "tags": ["Users"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Akun pengguna berhasil dihapus.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token akses tidak valid atau tidak diberikan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Pengguna tidak ditemukan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/update_user": {
      "put": {
        "summary": "Update Informasi Pengguna",
        "tags": ["Users"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Data pengguna yang akan diperbarui",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Informasi pengguna berhasil diperbarui.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Data tidak valid atau konflik data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token akses tidak valid atau tidak diberikan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Pengguna tidak ditemukan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/stocks/add": {
      "post": {
        "summary": "Tambah atau Update Stok Barang",
        "tags": ["Stocks"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Data stok barang yang akan ditambahkan",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockAdd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stok barang berhasil ditambahkan atau diupdate.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockResponse"
                }
              }
            }
          },
          "400": {
            "description": "Data tidak valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token akses tidak valid atau tidak diberikan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/stocks/remove": {
      "post": {
        "summary": "Kurangi Stok Barang",
        "tags": ["Stocks"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Data stok barang yang akan dikurangi",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockRemove"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stok barang berhasil dikurangi.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockResponse"
                }
              }
            }
          },
          "400": {
            "description": "Data tidak valid atau stok tidak mencukupi.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token akses tidak valid atau tidak diberikan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Barang tidak ditemukan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/stocks/check": {
      "get": {
        "summary": "Cek Semua Stok Barang",
        "tags": ["Stocks"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar semua stok barang.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stocks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/StockResponse"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token akses tidak valid atau tidak diberikan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/stocks/delete": {
      "delete": {
        "summary": "Hapus Stok Barang",
        "tags": ["Stocks"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Data barang yang akan dihapus",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockDelete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stok barang berhasil dihapus.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Data tidak valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token akses tidak valid atau tidak diberikan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Barang tidak ditemukan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Users",
      "description": "Endpoint untuk manajemen pengguna"
    },
    {
      "name": "Stocks",
      "description": "Endpoint untuk manajemen stok barang"
    }
  ]
}